Fix Salome features detection when lmod environment is present.
The lmod macros work in a bash environement, but not in a simple
Bourne shell, so it is necessary to force the bash environment for
subshells to avoid subshell errors. As a reminder, we use tests in subshells
to only load the Salome environment temporarily.

We know of no current architecture in which both Salome and code_saturne
can be compiled natively that does not include a bash shell, so this
does not cause any additional portability issue.

A cleaner way of detecting the Salome environment would be to use the
newer Salome  Python and .cfg file-based configuration.